Remove duplication in database
By defining the cmd including auth, the definitions are simplified.

It also tightens the data types.

+++ manifests/database.pp
@@ -1,51 +1,38 @@
 # == Class: influxdb::database
 #
 define influxdb::database (
-  Enum['absent', 'present'] $ensure  = present,
-  $db_name                           = $title,
-  $https_enable                      = $influxdb::https_enable,
-  $http_auth_enabled                 = $influxdb::http_auth_enabled,
-  $admin_username                    = $influxdb::admin_username,
-  $admin_password                    = $influxdb::admin_password
+  Enum['absent', 'present'] $ensure   = present,
+  String[1] $db_name                  = $title,
+  Boolean $https_enable               = $influxdb::https_enable,
+  Boolean $http_auth_enabled          = $influxdb::http_auth_enabled,
+  Optional[String[1]] $admin_username = $influxdb::admin_username,
+  Optional[String[1]] $admin_password = $influxdb::admin_password
 ) {
   if $https_enable {
-    $cmd = 'influx -ssl -unsafeSsl'
+    $args_https = ' -ssl -unsafeSsl'
   } else {
-    $cmd = 'influx'
+    $args_https = ''
   }
-  if ($ensure == 'absent') and ($http_auth_enabled == true) {
-    exec { "drop_database_${db_name}":
-      path    => '/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin',
-      command =>
-        "${cmd} -username ${admin_username} -password '${admin_password}' \
-        -execute 'DROP DATABASE ${db_name}'",
-      onlyif  =>
-        "${cmd} -username ${admin_username} -password '${admin_password}' \
-        -execute 'SHOW DATABASES' | tail -n+3 | grep -x ${db_name}"
-    }
-  } elsif ($ensure == 'present') and ($http_auth_enabled == true) {
-    exec { "create_database_${db_name}":
-      path    => '/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin',
-      command =>
-        "${cmd} -username ${admin_username} -password '${admin_password}' \
-        -execute 'CREATE DATABASE ${db_name}'",
-      unless  =>
-        "${cmd} -username ${admin_username} -password '${admin_password}' \
-        -execute 'SHOW DATABASES' | tail -n+3 | grep -x ${db_name}"
-    }
-  } elsif ($ensure == 'present') and ($http_auth_enabled == false) {
-    exec { "create_database_${db_name}":
-      path    => '/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin',
-      command => "${cmd} -execute 'CREATE DATABASE ${db_name}'",
-      unless  =>
-        "${cmd} -execute 'SHOW DATABASES' | tail -n+3 | grep -x ${db_name}"
-    }
-  } elsif ($ensure == 'absent') and ($http_auth_enabled == false) {
+
+  if $http_auth_enabled {
+    $args_auth = " -username ${admin_username} -password '${admin_password}'"
+  } else {
+    $args_auth = ''
+  }
+
+  $cmd = "influx${args_https}${args_auth}"
+
+  if $ensure == 'absent' {
     exec { "drop_database_${db_name}":
       path    => '/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin',
       command => "${cmd} -execute 'DROP DATABASE ${db_name}'",
-      onlyif  =>
-        "${cmd} -execute 'SHOW DATABASES' | tail -n+3 | grep -x ${db_name}"
+      onlyif  => "${cmd} -execute 'SHOW DATABASES' | tail -n+3 | grep -x ${db_name}",
+    }
+  } elsif $ensure == 'present' {
+    exec { "create_database_${db_name}":
+      path    => '/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin',
+      command => "${cmd} -execute 'CREATE DATABASE ${db_name}'",
+      unless  => "${cmd} -execute 'SHOW DATABASES' | tail -n+3 | grep -x ${db_name}",
     }
   }
 }
